(refactor): extract compile bench scenarios into common module
Move BenchConfig, build_db, and the per-phase setup blocks from compile.rs
into tests/benches/common.rs, exposing each phase as a one-shot run_* helper.
compile.rs becomes a thin criterion wrapper. No behavior change.
